LayoutTests/crypto/subtle/pbkdf2-derive-bits-malformed-parametrs-expected.txt

Test PBKDF2 deriveBits operation with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.deriveBits("PBKDF2", baseKey, null) rejected promise  with TypeError: Member Pbkdf2Params.hash is required and must be an instance of any.
PASS crypto.subtle.deriveBits({name: "PBKDF2"}, baseKey, null) rejected promise  with TypeError: Member Pbkdf2Params.hash is required and must be an instance of any.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: true, iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: null, iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: undefined, iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Member Pbkdf2Params.salt is required and must be an instance of BufferSource.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: Symbol(), iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: { }, iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: 1, iterations: 100000, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Type error.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: undefined, hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Member Pbkdf2Params.iterations is required and must be an instance of unsigned long.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: Symbol(), hash: "sha-1"}, baseKey, 128) rejected promise  with TypeError: Cannot convert a symbol to a number.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: 100000, hash: "sha-1"}, baseKey, null) rejected promise  with OperationError (DOM Exception 34): The operation failed for an operation-specific reason.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: 100000, hash: "sha-1"}, baseKey, 0) rejected promise  with OperationError (DOM Exception 34): The operation failed for an operation-specific reason.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: 100000, hash: "sha-1"}, baseKey, 5) rejected promise  with OperationError (DOM Exception 34): The operation failed for an operation-specific reason.
PASS crypto.subtle.deriveBits({name: "PBKDF2", salt: salt, iterations: 0, hash: "sha-1"}, baseKey, 128) rejected promise  with OperationError (DOM Exception 34): The operation failed for an operation-specific reason.
PASS successfullyParsed is true

TEST COMPLETE
